+++ test/Spec.hs
@@ -2,6 +2,8 @@ module Main (main) where
 
 import qualified Canon.Antlr4.CommentTest
 import qualified Canon.Antlr4.EscapeTest
+import qualified Canon.Antlr4.QueryTest
+import qualified Canon.Antlr4.RuleGraphTest
 import qualified Canon.Antlr4.RoundTripTest
 import qualified Canon.Antlr4.VendoredTest
 import qualified CanonTest
@@ -18,6 +20,8 @@ main =
             [ Canon.Antlr4.EscapeTest.tests
             , Canon.Antlr4.CommentTest.tests
             , Canon.Antlr4.RoundTripTest.tests
+            , Canon.Antlr4.QueryTest.tests
+            , Canon.Antlr4.RuleGraphTest.tests
             ]
         ]
     )